analysis/requirement.txt
Django==4.2.2
python-twitter==3.5
nltk==3.8.1
tweepy==4.14.0
